feat(features): AI 챗봇 fetch 직접 호출 aiClient로 추상화
features/ai-chat/api/aiClient.ts에 sendChatMessage 추출
AIChat 페이지에서 직접 fetch 호출 제거

closes #8
